use long for power levels

+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/user/UserService.kt
@@ -37,12 +37,12 @@ interface UserService {
 
     fun getById(roomId: RoomId, userId: UserId): Flow<RoomUser?>
 
-    fun getPowerLevel(roomId: RoomId, userId: UserId): Flow<Int>
+    fun getPowerLevel(roomId: RoomId, userId: UserId): Flow<Long>
     fun getPowerLevel(
         userId: UserId,
         powerLevelsEventContent: PowerLevelsEventContent?,
         createEventContent: CreateEventContent?
-    ): Int
+    ): Long
 
     fun canKickUser(roomId: RoomId, userId: UserId): Flow<Boolean>
     fun canBanUser(roomId: RoomId, userId: UserId): Flow<Boolean>
@@ -51,7 +51,7 @@ interface UserService {
     fun canInvite(roomId: RoomId): Flow<Boolean>
     fun canRedactEvent(roomId: RoomId, eventId: EventId): Flow<Boolean>
 
-    fun canSetPowerLevelToMax(roomId: RoomId, userId: UserId): Flow<Int?>
+    fun canSetPowerLevelToMax(roomId: RoomId, userId: UserId): Flow<Long?>
 
     @Deprecated("use canSendEvent instead", ReplaceWith("canSendEvent(roomId, RoomMessageEventContent::class)"))
     fun canSendMessages(roomId: RoomId): Flow<Boolean>
@@ -126,7 +126,7 @@ class UserServiceImpl(
     override fun getPowerLevel(
         roomId: RoomId,
         userId: UserId
-    ): Flow<Int> =
+    ): Flow<Long> =
         combine(
             roomStateStore.getContentByStateKey<PowerLevelsEventContent>(roomId),
             roomStateStore.getContentByStateKey<CreateEventContent>(roomId)
@@ -138,7 +138,7 @@ class UserServiceImpl(
         userId: UserId,
         powerLevelsEventContent: PowerLevelsEventContent?,
         createEventContent: CreateEventContent?
-    ): Int {
+    ): Long {
         return when (powerLevelsEventContent) {
             null -> if (createEventContent?.creator == userId) 100 else 0
             else -> powerLevelsEventContent.users[userId] ?: powerLevelsEventContent.usersDefault
@@ -258,7 +258,7 @@ class UserServiceImpl(
     override fun canSetPowerLevelToMax(
         roomId: RoomId,
         userId: UserId
-    ): Flow<Int?> {
+    ): Flow<Long?> {
         return combine(
             getPowerLevel(roomId, ownUserId),
             getPowerLevel(roomId, userId),
